chore: refactoring provision packages
