+++ extensions/provision-ionos-s3/src/main/java/com/ionos/edc/provision/s3/IonosProvisionExtension.java
@@ -16,10 +16,9 @@ package com.ionos.edc.provision.s3;
 
 import com.ionos.edc.extension.s3.connector.S3Connector;
 import com.ionos.edc.extension.s3.types.IonosToken;
-import com.ionos.edc.provision.s3.bucket.IonosS3ConsumerResourceDefinitionGenerator;
-import com.ionos.edc.provision.s3.bucket.IonosS3ProvisionedResource;
-import com.ionos.edc.provision.s3.bucket.IonosS3Provisioner;
-import com.ionos.edc.provision.s3.bucket.IonosS3ResourceDefinition;
+import com.ionos.edc.provision.s3.resource.IonosS3ConsumerResourceDefinitionGenerator;
+import com.ionos.edc.provision.s3.resource.IonosS3ProvisionedResource;
+import com.ionos.edc.provision.s3.resource.IonosS3ResourceDefinition;
 import dev.failsafe.RetryPolicy;
 import org.eclipse.edc.connector.controlplane.transfer.spi.provision.ProvisionManager;
 import org.eclipse.edc.connector.controlplane.transfer.spi.provision.ResourceManifestGenerator;
